2.5 add to lint ineffassign, intrange, modernize, tparallel, unconvert, usestdlibvars, usetesting (#5573)

+++ core/chaincode/handler.go
@@ -982,7 +982,7 @@ func getQueryMetadataFromBytes(metadataBytes []byte) (*pb.QueryMetadata, error) 
 func (h *Handler) calculateTotalReturnLimit(metadata *pb.QueryMetadata) int32 {
 	totalReturnLimit := int32(h.TotalQueryLimit)
 	if metadata != nil {
-		pageSize := int32(metadata.GetPageSize())
+		pageSize := metadata.GetPageSize()
 		if pageSize > 0 && pageSize < totalReturnLimit {
 			totalReturnLimit = pageSize
 		}
